insert __init__.py patch after __future__ imports
__future__ imports must occur near the beginning of __init__.py to be
syntactically valid. This change ensures that when __init__.py is
patched, the patch is inserted after (rather than before) any __future__
imports. Fixes #8.

One limitation of this change is that we do not handle __future__ import
statements that are split across multiple lines. For example, if
__init__.py contains 'from __future__ \\\n    import print_function',
an exception will be raised during the wheel repair step.

Another limitation is that we do not handle the case where an additional
statement is on the same line as a __future__ import, and this second
statement requires the patch to work properly. For example, if
__init__.py contains 'from __future__ import print_function; import
my_extension', then the patch will be inserted after this line, even if
the statement 'import my_extension' requires the patch to work.

However, these 2 limitations should not be a problem in the vast
majority of cases, as these are unusual edge cases that are unlikely to
be found in the __init__.py files in the wild. Even if these cases do
exist, it is simple for the user to change __init__.py to avoid these
cases.