Fix DLL load error for Anaconda Python 3.8 and 3.9
Further testing on Anaconda distributions of Python have revealed that
the previous workaround with CONDA_DLL_SEARCH_MODIFICATION_ENABLE does
not work as expected, and the following error persists.

ImportError: DLL load failed while importing error: The specified module
could not be found.

Anaconda has fixed the bug regarding os.add_dll_directory() starting in
Python 3.10. Thus, we will now use that function for Anaconda Python as
long as the version is at least 3.10. For Anaconda Python < 3.10, we go
back to the strategy of preloading the DLLs.