Revise lists of system DLLs
Previously, the lists of system DLLs to avoid vendoring were obtained
from the System32 and SysWOW64 directories of a Windows 7 SP1 instance.
The lists have been revised to include only the DLLs that are present in
Windows 7 SP1, 8, 8.1, 10, and 11.

This also fixes a mistake where the list of 64-bit system DLLs was
incorrect; it was erroneously copied from the list of 32-bit system
DLLs.