Don't fail fast in __init__.py patch for PyInstaller environment
When PyInstaller is used, files and folders expected in the __init__.py
patch might not exist. To avoid raising an error immediately, we first
check for existence of files and folders before attempting to access
them if we are in a PyInstaller environment.

In such a PyInstaller environment, it is possible that a DLL load error
will happen later. However, we avoid raising an error if PyInstaller
placed those DLLs in a different location in the DLL search path or if
no dependent extension module is imported. Ultimately, it will be
PyInstaller's responsibility to ensure that dependent DLLs are found at
runtime.

See https://github.com/pyinstaller/pyinstaller-hooks-contrib/issues/421
for more details.