Remove Visual C++ 14.x runtime redistributable directory from default search path
Previously, we would try to find this directory when searching for DLLs,
but it was highly environment-dependent whether we would find this
directory successfully. setuptools needed to be installed, and it would
need to be successful in finding this directory. The search for this
directory could fail in multiple ways, and we would silently ignore all
errors.

To maintain better consistency in how delvewheel works in different
environments, we no longer try to add this directory to the default
search path. I expect this to have little impact for most users, and
those who are affected can easily add this directory to the PATH
environment variable or use the --add-path option.