Improve Anaconda Python detection
Previously, the presence of the conda-meta directory was used to
determine whether Anaconda Python was used. This directory might not
exist if the Python distribution is repackaged (with PyInstaller for
example). Instead, we now check for the existence of the
Anaconda_GetVersion() function in the C API, which is present in the
main channel Python distributions, or the string 'packaged by
conda-forge', which is present in the conda-forge channel distributions.

In addition, we add an explicit check for the CPython implementation
because the conda-forge PyPy distribution does not have the
os.add_dll_directory() bug.

Not all Anaconda Python versions can be detected this way, but the
distinction between Anaconda Python and other distributions matters only
for versions 3.8 and 3.9. I checked that all Anaconda Python
distributions in this version range satisfy this check.