Use wheel distribution name in hash for name-mangling
Previously, the binary contents of a DLL prior to the name-mangling step
was used to compute its hash. This left a rare edge case where the wrong
DLL could be loaded if the following were all true.
- Two wheels are installed and used at the same time in a Python
  process.
- There exists a DLL that is vendored by both wheels, and this DLL was
  byte-for-byte identical between the two wheels prior to name-mangling.
  Let example.dll denote this wheel prior to name-mangling.
- Between the two wheels, the dependencies of example.dll had the same
  name but are not cross-compatible. That is, the first wheel will
  encounter issues if it uses a transitive dependency from the second
  wheel.
To avoid this situation, use both the wheel distribution name and the
DLL contents when computing a hash. This maintains the advantage of
deterministic, repeatable hashes for a given distribution.