Use internal padding space when mangling DLL names
Previously, we were unable to mangle a DLL name if a dependent DLL
contained unstrippable trailing data at the end of the binary. Trailing
data prevented us from name-mangling because we were unable to add a new
PE section to the end of the binary containing the mangled names.

Now, we first try to find any internal padding space within a DLL to fit
the mangled names. If enough padding exists, then we don't need to add a
new PE section and thus no longer need to care whether trailing data
exist. Of course, if not enough internal padding exists, then we fall
back to the previous strategy of creating a new PE section.

This strategy reduces the chance of errors in the name-mangling step if
a DLL contains trailing data. This strategy also has the benefit of not
increasing DLL file size if enough internal padding exists.