Remove load order calculation
Previously, we used a topological sort to determine the order in which
to preload DLLs for Python < 3.8 or Anaconda Python < 3.10. Now, we call
the Windows function LoadLibraryExW() directly with the
LOAD_WITH_ALTERED_SEARCH_PATH flag when preloading, so no sorting is
required anymore.

We still create and use a .load-order file to maintain backward
compatibility with re-bundling tools such as PyInstaller and to avoid
automatically loading additional DLLs added with --add-dll.