Revert "Improve support for delay-load DLLs"
This reverts commit 2328c3718a16ead5def934d9be84ef389415fb17.
Adding support for delay-loaded DLLs has been found to cause other
issues such as #34. Another issue I found is that if we preload the
DLLs at Python startup, then pip uninstall will fail because the DLLs
are locked by the Python process running pip.

Given that it seems that delay-load DLLs aren't used too frequently
anyway, I am a backing out this change. If there is demand in the future
for better support for delay-load DLLs, we can reassess.