Provide delay-load import hook
Projects that have a delay-load DLL dependency will need to build a
custom hook into the DLL which has such a dependency in order to ensure
that the directory containing the vendored DLLs is in the DLL search path
when delay-loading. Provide the C source code of a hook that
accomplishes this.