Introduce support for namespace packages
The --namespace-pkg flag can be used to specify which packages are
namespace packages. We do not create or modify __init__.py if it is in a
namespace package. To ensure that vendored DLLs can still be loaded at
runtime, we do one or more of the following depending on the situation.
- Patch the wheel one directory level deeper.
- Patch a .py file.
- Include an extra copy of the DLLs.

Fix #38