Introduce the `--analyze-existing` option
When specified, this option causes delvewheel to analyze and vendor in
dependencies of DLLs that are already in the wheel. The intention is to
support the use case where a wheel already contains a DLL which is
loaded using LoadLibrary(), and this DLL has dependencies that should be
vendored and name-mangled.

It is the responsibility of the application to ensure that the directory
containing the vendored DLLs is in the DLL search path when
LoadLibrary() is called.

Fix #47