Clear the `/DEPENDENTLOADFLAG` value
If a vendored DLL was built with a non-0 value for `/DEPENDENTLOADFLAG`,
then it can fail to load because it changes the DLL search path for its
dependencies. Modify any vendored DLLs (other than those specified with
`--add-dll`) with non-0 `/DEPENDENTLOADFLAG` to clear the value to 0.
This ensures that the default search path for Python extension modules
is used.

Fix #49