Fix `OSError: Error loading *.dll; The operation completed successfully.`
When LoadLibraryExW() is called from `ctypes`, the 64-bit void* return
value is cast to a 32-bit C int by default. If the base address of the
loaded DLL is a multiple of 2**32, then the lower 32 bits are 0, so
Python sees a return value of 0 and assumes that the DLL failed to load.
Set the return type properly so that the downcast is not done.

This fixes an intermittent issue with Python <= 3.7 or Conda Python <=
3.9 where an `OSError` is raised saying that a DLL failed to load but
also that the operation completed successfully.

Fix #51