Fix overzealous warnings if vendored Microsoft Visual C++ runtime is too old
Previously, a warning was issued if the Microsoft Visual C++ runtime was
only slightly older than the version that the application was built
against. The warning was too sensitive because the Microsoft Visual C++
runtime might still be new enough to provide compatibility. Now, issue a
warning only if the Microsoft Visual C++ runtime comes from an older
platform toolset than that which the application was built against.