Revise hashing algorithm for name mangling to permit limited DLL sharing between wheels
Previously during name-mangling, the hash appended to a DLL was derived
from the DLL contents prior to name mangling along with the wheel
distribution name. This ensured that name-mangled DLLs were never shared
between projects. However, DLLs should be safe to share between projects
if they are bitwise identical prior to name-mangling and all their
dependencies that will be name-mangled are also bitwise identical.

To allow for this limited sharing, change the hashing algorithm to
derive the hash from the DLL contents and the hashes of all its direct
dependencies which will be name-mangled. This saves memory at DLL load
time by ensuring just one copy of the DLL is loaded when it is safe for
two projects to use the same copy.

One beneficial effect of this change is that if two projects use the
same exact version of msvcp140.dll, then they will attach the same hash
to their vendored copies of msvcp140.dll. If both projects are used in
the same python.exe process, then only a single copy of the DLL is
loaded.