Fix `UnicodeDecodeError` when patching `.py` files containing Unicode characters
The default file encoding on Windows is Windows-1252, while Python
source files are typically UTF-8-encoded. To avoid encoding/decoding
errors, two strategies are now used.

One is to read/write files in UTF-8 mode with the `'surrogateescape'`
error handler. This can handle source files with any encoding.

The other is to read/write in binary mode, where the encoding does not
matter.