Improve the MSVC++ redistributable version check warning
The previous check failed to handle the case that an application using
`std::mutex` which links against a 14.4x redistributable is not
compatible with a 14.3x redistributable, even though both versions use
the v143 platform toolset. Change the version check to consider up to
the tens digit of the redistributable version.